+++ src/main/kotlin/lotto/Main.kt
@@ -1,5 +1,15 @@
 package lotto
 
+import lotto.domain.LottoFactory
+import lotto.domain.LottoIncomeCalculator
+import lotto.domain.LottoNumber
+import lotto.domain.LottoService
+import lotto.domain.LottoStore
+import lotto.domain.Money
+import lotto.domain.RandomNumberGenerator
+import lotto.view.ConsoleInputView
+import lotto.view.ConsoleOutputView
+
 fun main() {
     val inputView = ConsoleInputView()
     val outputView = ConsoleOutputView()
