+++ src/main/kotlin/lotto/domain/IncomeCalculator.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 fun interface IncomeCalculator {
     fun calculate(
