+++ src/main/kotlin/lotto/domain/IncomeRate.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 @JvmInline
 value class IncomeRate(val value: Double)
