+++ src/main/kotlin/lotto/domain/LottoService.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 class LottoService(
     private val lottoStore: LottoStore,
