+++ src/main/kotlin/lotto/domain/LottoUtil.kt
@@ -1,3 +1,3 @@
-package lotto
+package lotto.domain
 
 const val LOTTO_PRICE = 1000
