+++ src/main/kotlin/lotto/domain/Lottos.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 data class Lottos(
     val lottos: List<Lotto>,
