+++ src/main/kotlin/lotto/domain/Result.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 enum class Result(
     private val predicate: (Int) -> Boolean,
