+++ src/main/kotlin/lotto/view/InputView.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.view
 
 interface InputView {
     fun requestPrice(): Int
