+++ src/test/kotlin/lotto/fixture/LottoFixture.kt
@@ -1,6 +1,6 @@
 package lotto.fixture
 
-import lotto.Lotto
-import lotto.LottoNumber
+import lotto.domain.Lotto
+import lotto.domain.LottoNumber
 
 internal fun Lotto.Companion.fixture(vararg numbers: Int): Lotto = Lotto(numbers.map(::LottoNumber))
