+++ src/test/kotlin/lotto/fixture/LottoNumberFixture.kt
@@ -1,5 +1,5 @@
 package lotto.fixture
 
-import lotto.LottoNumber
+import lotto.domain.LottoNumber
 
 internal fun lottoNumbersFixture(vararg numbers: Int): Set<LottoNumber> = numbers.map(::LottoNumber).toSet()
